style: rename data to cuddle
